Move Overload to VariantUtils
